AI-Orchestrated-Headless-Composable-Application-Platform-main/telegram_client/requirements.txt
aiogram>=3.4.1
httpx>=0.27.0
python-dotenv>=1.0.1
